+++ modules/polyfills/src/images-node/parse-image.node.js
@@ -10,9 +10,17 @@ export async function parseImageNode(arrayBuffer, mimeType, options) {
   const buffer = arrayBuffer instanceof Buffer ? arrayBuffer : Buffer.from(arrayBuffer);
 
   const ndarray = await getPixelsAsync(buffer, mimeType);
+
+  const shape = [...ndarray.shape];
+  const layers = ndarray.shape.length === 4 ? ndarray.shape.shift() : 1;
+
+  // extract width/height etc
   return {
+    shape,
+    data: ndarray.data,
     width: ndarray.shape[0],
     height: ndarray.shape[1],
-    data: ndarray.data
+    components: ndarray.shape[2],
+    layers
   };
 }